Create basic controllers

+++ Web/Backend/AttendanceManager/AttendanceManager.Core/Interfaces/DAL/IRepository.cs
@@ -6,7 +6,7 @@ namespace AttendanceManager.Core.Interfaces.DAL
 {
     public interface IRepository<TModel> where TModel : class
     {
-        TModel GetById(long id);
+        TModel GetById(int id);
         IQueryable<TModel> GetAll();
         IQueryable<TModel> Query(Expression<Func<TModel, bool>> predicate);
         TModel Add(TModel model);
